+++ src/main/java/com/example/myapp/dto/Cart_dto.java
@@ -1,5 +1,7 @@
 package com.example.myapp.dto;
 
+import java.math.BigDecimal;
+
 import com.example.myapp.data.Book;
 
 import lombok.Data;
@@ -9,12 +11,12 @@ public class Cart_dto {
 
     private int bookId ; 
 
-    private double bookPrice ; 
+    private BigDecimal bookPrice ; 
 
     private String bookName ;
 
     private String cover ; 
-    public Cart_dto(int userId, int bookId , double bookPrice , String bookName , String cover)
+    public Cart_dto(int userId, int bookId , BigDecimal bookPrice , String bookName , String cover)
     {
         this.bookId = bookId ; 
         this.bookPrice = bookPrice ;
